+++ lib/librte_port/rte_port_crypto.c
@@ -213,7 +213,7 @@ struct rte_port_crypto_writer {
 	uint16_t qp_id;
 
 	uint16_t block_size;
-	unsigned digest_length;
+	uint16_t digest_length;
 	uint32_t burst_sz;
 //	uint64_t bsz_mask
 
@@ -377,7 +377,7 @@ rte_port_crypto_writer_tx(void *port, struct rte_mbuf *pkt)
 
 	printf("######## crypto_tx: do_hash=%d, do_cipher=%d, hash_verify=%d\n",
 			p->do_hash, p->do_cipher, p->hash_verify);
-	printf("######## crypto_tx: data_len=%d, p->digest_length=%d\n",
+	printf("######## crypto_tx: data_len=%d, p->digest_length=%u\n",
 					data_len, p->digest_length);
 	if (p->do_hash) {
 		if (!p->hash_verify) {
